requirements.txt
sqlalchemy~=2.0.20
psycopg2-binary
python-dotenv~=1.0.0
pytest~=7.4.0
fastapi~=0.103.0
uvicorn~=0.23.2
pydantic~=2.3.0
PyJWT~=2.8.0
passlib~=1.7.4